+++ src/presentation/controllers/signup.spec.ts
@@ -1,6 +1,5 @@
 import { throws } from "assert";
-import { InvalidParameterError } from "../errors/invalid-params-error";
-import { MissingParameterError } from "../errors/missing-params-error";
+import { InvalidParameterError, MissingParameterError } from "../errors";
 import { ServerError } from "../errors/server-error";
 import { IEmailValidator } from "../protocols/email-validation";
 import { SignUpController } from "./signup";
